[BE] Refactor: 리뷰 상인 평균평점 반영 로직 추가 (#277)
* feat: 리뷰 작성시, 상인 평균평점에 반영 로직 추가

* Refactor: 메서드에서 static 키워드 제거

* Refactor: 평균평점 구하는 메서드 Member도메인으로 이동

+++ backend/src/main/java/com/dongnebook/domain/member/dto/response/BookDetailMemberResponse.java
@@ -8,11 +8,11 @@ import lombok.Getter;
 public class BookDetailMemberResponse {
 	private Long merchantId;
 	private String name;
-	private Long grade;
+	private Double grade;
 	private String avatarUrl;
 
 	@QueryProjection
-	public BookDetailMemberResponse(Long merchantId, String name, Long grade, String avatarUrl) {
+	public BookDetailMemberResponse(Long merchantId, String name, Double grade, String avatarUrl) {
 		this.merchantId = merchantId;
 		this.name = name;
 		this.grade = grade;
